% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto.R
\name{carto_styles}
\alias{carto_styles}
\alias{carto_voyager}
\alias{carto_positron}
\alias{carto_dark_matter}
\alias{carto_voyager_nolabels}
\alias{carto_positron_nolabels}
\alias{carto_dark_matter_nolabels}
\title{Carto styles}
\usage{
carto_voyager()

carto_positron()

carto_dark_matter()

carto_voyager_nolabels()

carto_positron_nolabels()

carto_dark_matter_nolabels()
}
\description{
Carto style URLs.
}
\section{Functions}{
\itemize{
\item \code{carto_voyager()}: Carto Voyager

\item \code{carto_positron()}: Carto Positron

\item \code{carto_dark_matter()}: Carto Dark Matter

\item \code{carto_voyager_nolabels()}: Carto Voyager (no labels)

\item \code{carto_positron_nolabels()}: Carto Positron (no labels)

\item \code{carto_dark_matter_nolabels()}: Carto Dark Matter (no labels)

}}
\seealso{
\url{https://github.com/CartoDB/basemap-styles}
}
