% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_studio.R
\name{mapbox_studio_styles}
\alias{mapbox_studio_styles}
\alias{mapbox_studio_galaxy}
\title{Mapbox studio styles}
\usage{
mapbox_studio_galaxy()
}
\description{
Mapbox studio style URLs.
}
\section{Functions}{
\itemize{
\item \code{mapbox_studio_galaxy()}: Mapbox Studio Galaxy

}}
\seealso{
\url{https://studio.mapbox.com/}
}
