% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uber.R
\name{uber_styles}
\alias{uber_styles}
\alias{uber_dark}
\alias{uber_light}
\alias{uber_muted_light}
\alias{uber_muted_night}
\title{Mapbox gallery styles}
\usage{
uber_dark()

uber_light()

uber_muted_light()

uber_muted_night()
}
\description{
Mapbox gallery style URLs.
}
\section{Functions}{
\itemize{
\item \code{uber_dark}: Uber Dark

\item \code{uber_light}: Uber Light

\item \code{uber_muted_light}: Uber Muted Light

\item \code{uber_muted_night}: Uber Muted Night
}}

\seealso{
\url{https://github.com/keplergl/kepler.gl/blob/v2.3.0/src/constants/default-settings.js#L228-L263}
}
