% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{mapbox_styles}
\alias{mapbox_styles}
\alias{mapbox_streets}
\alias{mapbox_outdoors}
\alias{mapbox_light}
\alias{mapbox_dark}
\alias{mapbox_satellite}
\alias{mapbox_satellite_streets}
\alias{mapbox_navigation_day}
\alias{mapbox_navigation_night}
\alias{mapbox_navigation_guidance_night}
\alias{mapbox_navigation_guidance_day}
\title{Mapbox styles}
\usage{
mapbox_streets(version = 12)

mapbox_outdoors(version = 12)

mapbox_light(version = 10)

mapbox_dark(version = 10)

mapbox_satellite(version = 9)

mapbox_satellite_streets(version = 12)

mapbox_navigation_day(version = 1)

mapbox_navigation_night(version = 1)

mapbox_navigation_guidance_night(version = 4)

mapbox_navigation_guidance_day(version = 4)
}
\arguments{
\item{version}{\verb{<integer>} The style version number}
}
\description{
Mapbox style URLs.
}
\section{Functions}{
\itemize{
\item \code{mapbox_streets()}: Mapbox Streets

\item \code{mapbox_outdoors()}: Mapbox Outdoors

\item \code{mapbox_light()}: Mapbox Light

\item \code{mapbox_dark()}: Mapbox Dark

\item \code{mapbox_satellite()}: Mapbox Satellite

\item \code{mapbox_satellite_streets()}: Mapbox Satellite Streets

\item \code{mapbox_navigation_day()}: Mapbox Navigation Day

\item \code{mapbox_navigation_night()}: Mapbox Navigation Night

\item \code{mapbox_navigation_guidance_night()}: Navigation Guidance Night

\item \code{mapbox_navigation_guidance_day()}: Navigation Guidance Day

}}
\seealso{
\url{https://docs.mapbox.com/api/maps/#mapbox-styles}
}
