% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{mapbox_styles}
\alias{mapbox_styles}
\alias{mapbox_streets}
\alias{mapbox_outdoors}
\alias{mapbox_light}
\alias{mapbox_dark}
\alias{mapbox_satellite}
\alias{mapbox_satellite_streets}
\alias{mapbox_navigation_day}
\alias{mapbox_navigation_night}
\alias{mapbox_navigation_guidance_night}
\alias{mapbox_navigation_guidance_day}
\title{Mapbox styles}
\usage{
mapbox_streets()

mapbox_outdoors()

mapbox_light()

mapbox_dark()

mapbox_satellite()

mapbox_satellite_streets()

mapbox_navigation_day()

mapbox_navigation_night()

mapbox_navigation_guidance_night()

mapbox_navigation_guidance_day()
}
\description{
Mapbox style URLs.
}
\section{Functions}{
\itemize{
\item \code{mapbox_streets}: Mapbox Streets

\item \code{mapbox_outdoors}: Mapbox Outdoors

\item \code{mapbox_light}: Mapbox Light

\item \code{mapbox_dark}: Mapbox Dark

\item \code{mapbox_satellite}: Mapbox Satellite

\item \code{mapbox_satellite_streets}: Mapbox Satellite Streets

\item \code{mapbox_navigation_day}: Mapbox Navigation Day

\item \code{mapbox_navigation_night}: Mapbox Navigation Night

\item \code{mapbox_navigation_guidance_night}: Navigation Guidance Night

\item \code{mapbox_navigation_guidance_day}: Navigation Guidance Day
}}

\seealso{
\url{https://docs.mapbox.com/api/maps/#mapbox-styles}
}
